# specs/001-oci-registry-server/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: jreg - OCI Distribution Spec v1.1.1
  description: |
    OCI-compliant container registry server implementing the OCI Distribution Specification v1.1.1.
    
    This API enables pushing and pulling OCI images and artifacts using Docker CLI, ORAS, and other OCI-compliant tools.
    
    **Key Features:**
    - Full OCI Distribution Spec v1.1.1 compliance
    - Content-addressable storage (SHA-256/SHA-512)
    - Chunked blob uploads with resume support
    - Multi-architecture image support (OCI Image Index)
    - Referrers API for artifact linking (signatures, SBOMs)
    - Tag management and discovery
    
    **Authentication:** Deferred for MVP (all endpoints publicly accessible)
  version: 1.0.0
  contact:
    name: jreg Development Team
  license:
    name: MIT
servers:
  - url: http://localhost:8080/v2
    description: Local development server
  - url: https://registry.example.com/v2
    description: Production server
...
paths:
...
  /{name}/blobs/uploads/:
    parameters:
      - $ref: '#/components/parameters/RepositoryName'
    
    post:
      operationId: startBlobUpload
      tags: [uploads]
      summary: Start blob upload
      description: |
        Initiate a blob upload session (OCI Distribution Spec end-4a).
        
        **Monolithic upload:** Provide `?digest=` query parameter to upload in single request.
        **Chunked upload:** Omit digest to start a session, then use PATCH to upload chunks.
        **Mount blob:** Use `?mount=<digest>&from=<repository>` to copy blob from another repo.
      parameters:
        - name: digest
          in: query
          required: false
          schema:
            type: string
            pattern: '^(sha256|sha512):[a-f0-9]{64,128}$'
          description: For monolithic upload (single POST with full content)
        - name: mount
          in: query
          required: false
          schema:
            type: string
          description: Mount blob from another repository (blob digest)
        - name: from
          in: query
          required: false
          schema:
            type: string
          description: Source repository for mount operation
      requestBody:
        description: Blob content (for monolithic upload only)
        required: false
        content:
          application/octet-stream:
            schema:
              type: string
              format: binary
      responses:
        '201':
          description: Blob uploaded successfully (monolithic)
          headers:
            Location:
              schema:
                type: string
                example: /v2/library/ubuntu/blobs/sha256:abc123...
              description: URL to access the blob
            Docker-Content-Digest:
              schema:
                type: string
        '202':
          description: Upload session started (chunked) or mount successful
          headers:
            Location:
              schema:
                type: string
                example: /v2/library/ubuntu/blobs/uploads/uuid-1234
              description: Upload session URL
            Range:
              schema:
                type: string
                example: 0-0
              description: Byte range uploaded so far
            Docker-Upload-UUID:
              schema:
                type: string
                format: uuid
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          description: Source blob not found (mount operation)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # end-4b: Get upload session status
  /{name}/blobs/uploads/{uuid}:
    parameters:
      - $ref: '#/components/parameters/RepositoryName'
      - $ref: '#/components/parameters/UploadUUID'
    
    get:
      operationId: getUploadStatus
      tags: [uploads]
      summary: Get upload session status
      description: |
        Get the status of an upload session (OCI Distribution Spec end-4b).
        Returns the byte range uploaded so far.
      responses:
        '204':
          description: Upload session exists
          headers:
            Range:
              schema:
                type: string
                example: 0-1048575
              description: Byte range uploaded
            Docker-Upload-UUID:
              schema:
                type: string
                format: uuid
        '404':
          $ref: '#/components/responses/UploadNotFound'

    # end-4c: Upload blob chunk
    patch:
      operationId: uploadBlobChunk
      tags: [uploads]
      summary: Upload blob chunk
      description: |
        Upload a chunk of blob data (OCI Distribution Spec end-4c).
        Chunks must be uploaded sequentially (no gaps).
      parameters:
        - name: Content-Range
          in: header
          required: true
          schema:
            type: string
            example: 0-1048575
          description: Byte range of this chunk
      requestBody:
        description: Chunk data
        required: true
        content:
          application/octet-stream:
            schema:
              type: string
              format: binary
      responses:
        '202':
          description: Chunk accepted
          headers:
            Location:
              schema:
                type: string
              description: Upload session URL
            Range:
              schema:
                type: string
              description: Total bytes uploaded
            Docker-Upload-UUID:
              schema:
                type: string
                format: uuid
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/UploadNotFound'
        '416':
          description: Invalid content range (gap detected)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    # end-4d: Complete blob upload
    put:
      operationId: completeBlobUpload
      tags: [uploads]
      summary: Complete blob upload
      description: |
        Finalize the blob upload (OCI Distribution Spec end-4d).
        Verifies digest matches uploaded content.
      parameters:
        - name: digest
          in: query
          required: true
          schema:
            type: string
            pattern: '^(sha256|sha512):[a-f0-9]{64,128}$'
          description: Expected digest of complete blob
      requestBody:
        description: Final chunk data (optional if all chunks uploaded via PATCH)
        required: false
        content:
          application/octet-stream:
            schema:
              type: string
              format: binary
      responses:
        '201':
          description: Blob upload completed
          headers:
            Location:
              schema:
                type: string
              description: URL to access the blob
            Docker-Content-Digest:
              schema:
                type: string
        '400':
          description: Digest mismatch or invalid upload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          $ref: '#/components/responses/UploadNotFound'

    # end-4e: Cancel upload session
    delete:
      operationId: cancelBlobUpload
      tags: [uploads]
      summary: Cancel upload session
      description: |
        Cancel an upload session and discard uploaded data (OCI Distribution Spec end-4e).
      responses:
        '204':
          description: Upload session cancelled
        '404':
          $ref: '#/components/responses/UploadNotFound'

  # end-5: Manifest existence check
...
components:
  parameters:
    RepositoryName:
      name: name
      in: path
      required: true
      schema:
        type: string
        pattern: '^[a-z0-9]+((\.|_|__|-+)[a-z0-9]+)*(\/[a-z0-9]+((\.|_|__|-+)[a-z0-9]+)*)*$'
        example: library/ubuntu
      description: Repository name (lowercase, alphanumeric with separators)
...
    UploadUUID:
      name: uuid
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Upload session UUID
  schemas:
...
    ErrorResponse:
      type: object
      required:
        - errors
      properties:
        errors:
          type: array
          items:
            $ref: '#/components/schemas/Error'
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          enum:
            - BLOB_UNKNOWN
            - BLOB_UPLOAD_INVALID
            - BLOB_UPLOAD_UNKNOWN
            - DIGEST_INVALID
            - MANIFEST_BLOB_UNKNOWN
            - MANIFEST_INVALID
            - MANIFEST_UNKNOWN
            - NAME_INVALID
            - NAME_UNKNOWN
            - SIZE_INVALID
            - UNAUTHORIZED
            - DENIED
            - UNSUPPORTED
          description: OCI error code
        message:
          type: string
          description: Human-readable error message
        detail:
          type: object
          additionalProperties: true
          description: Additional error context
      example:
        code: BLOB_UNKNOWN
        message: Blob not found in repository
        detail:
          digest: sha256:abc123...
          repository: library/ubuntu
  responses:
    BadRequest:
      description: Bad request (invalid parameters)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    UploadNotFound:
      description: Upload session not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            errors:
              - code: BLOB_UPLOAD_UNKNOWN
                message: Upload session not found